- fix package path - added tests
