+++ src/main/java/com/example/coffeelogger/CoffeeLoggerApplication.java
@@ -1,0 +1,14 @@
+package com.example.coffeelogger;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CoffeeLoggerApplication {
+
+    public static void main(String[] args) {
+
+        SpringApplication.run(CoffeeLoggerApplication.class, args);
+
+    }
+}
