+++ src/main/java/com/example/coffeelogger/bot/longpolling/Bot.java
@@ -1,0 +1,46 @@
+package com.example.coffeelogger.bot.longpolling;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import org.telegram.telegrambots.bots.TelegramLongPollingBot;
+import org.telegram.telegrambots.meta.api.objects.Update;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+import com.example.coffeelogger.bot.longpolling.service.BotService;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class Bot extends TelegramLongPollingBot {
+    private final BotService botService;
+
+    @Override
+    public String getBotUsername() {
+        return "helper";
+    }
+
+    @Override
+    public String getBotToken() {
+        try {
+            return Files.readString(Path.of("bot_token.txt"));
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Override
+    public void onUpdateReceived(Update update) {
+        if (update.hasMessage()) {
+            var responseMessage = botService.handleMessage(update);
+            try {
+                this.execute(responseMessage);
+            } catch (TelegramApiException ex) {
+                log.error(ex.getMessage(), ex);
+            }
+        }
+    }
+}
