+++ src/main/java/com/example/coffeelogger/bot/longpolling/persistance/adapter/CoffeeOrderAdapter.java
@@ -1,0 +1,52 @@
+package com.example.coffeelogger.bot.longpolling.persistance.adapter;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import com.example.coffeelogger.bot.longpolling.persistance.repository.CoffeeOrderRepository;
+import com.example.coffeelogger.bot.longpolling.persistance.entity.Coffee;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Map;
+
+@Component
+@RequiredArgsConstructor
+public class CoffeeOrderAdapter {
+    private final CoffeeOrderRepository repository;
+
+    public void createOrder(Map<String, String> coffeeOrder, Long userId) {
+        var coffee = createCoffeeEntity(coffeeOrder, userId);
+        repository.save(coffee);
+    }
+
+    public List<Coffee> getWeekStats(Long userId) {
+        return repository.findCoffeesByUserIdAndCreateDateTimeAfter(userId, computeAfterDate(7L));
+    }
+
+    public List<Coffee> getMontStats(Long userId) {
+        return repository.findCoffeesByUserIdAndCreateDateTimeAfter(userId, computeAfterDate(31L));
+    }
+
+    public void repeatLastOrder(Long userId) {
+        var lastOrder = repository.findDistinctFirstByUserId(userId);
+        System.err.println(lastOrder.getCreateDateTime());
+    }
+
+    // ===================================================================================================================
+    // = Implementation
+    // ===================================================================================================================
+
+    private static Coffee createCoffeeEntity(Map<String, String> coffeeOrder, Long userId) {
+        var coffee = new Coffee();
+        coffee.setCafe(coffeeOrder.get("CafeName"));
+        coffee.setType(coffeeOrder.get("CoffeeType"));
+        coffee.setPrice(Integer.parseInt(coffeeOrder.get("CoffeePrise")));
+        coffee.setSize(Float.parseFloat(coffeeOrder.get("CoffeeSize")));
+        coffee.setUserId(userId);
+        return coffee;
+    }
+
+    private LocalDateTime computeAfterDate(Long daysCount) {
+        return LocalDateTime.now().minusDays(daysCount);
+    }
+}
