+++ src/main/java/com/example/coffeelogger/bot/longpolling/persistance/repository/CustomOrderRepository.java
@@ -1,0 +1,7 @@
+package com.example.coffeelogger.bot.longpolling.persistance.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import com.example.coffeelogger.bot.longpolling.persistance.entity.Custom;
+
+public interface CustomOrderRepository extends JpaRepository<Custom, String> {
+}
