+++ src/main/java/com/example/coffeelogger/bot/properties/BotProperties.java
@@ -1,0 +1,4 @@
+package com.example.coffeelogger.bot.properties;
+
+public class BotProperties {
+}
