--- src/main/java/ru/lanit/bpm/coffeelogger/CoffeeLoggerApplication.java
@@ -1,14 +1,0 @@
-package ru.lanit.bpm.coffeelogger;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CoffeeLoggerApplication {
-
-    public static void main(String[] args) {
-
-        SpringApplication.run(CoffeeLoggerApplication.class, args);
-
-    }
-}
