--- src/main/java/ru/lanit/bpm/coffeelogger/bot/longpolling/persistance/repository/CoffeeOrderRepository.java
@@ -1,11 +1,0 @@
-package ru.lanit.bpm.coffeelogger.bot.longpolling.persistance.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import ru.lanit.bpm.coffeelogger.bot.longpolling.persistance.entity.Coffee;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public interface CoffeeOrderRepository extends JpaRepository<Coffee, String> {
-    List<Coffee> findCoffeesByUserIdAndCreateDateTimeAfter(Long userId, LocalDateTime createDateTime);
-}
