--- src/main/java/ru/lanit/bpm/coffeelogger/bot/properties/BotProperties.java
@@ -1,4 +1,0 @@
-package ru.lanit.bpm.coffeelogger.bot.properties;
-
-public class BotProperties {
-}
